feat: merge /recentgames and /status
