+++ packages/discord/src/services/api.service.ts
@@ -16,7 +16,6 @@ import {
   GuildQuery,
   LeaderboardQuery,
   PlayerNotFoundException,
-  RecentGamesNotFoundException,
   SessionNotFoundException,
   ApiService as StatsifyApiService,
   StatusNotFoundException,
@@ -76,40 +75,6 @@ export class ApiService extends StatsifyApiService {
         throw new ErrorMessage(
           (t) => t("errors.invalidSession.title"),
           (t) => t("errors.invalidSession.description", { displayName: this.emojiDisplayName(t, displayName) })
-        );
-      }
-
-      throw this.unknownError();
-    });
-  }
-
-  /**
-   *
-   * @param tag Username, UUID, or Discord ID, or nothing. If nothing is provided it will attempt to fall back on the provided user.
-   * @param user User to use if no tag is provided.
-   * @returns The player's recent games
-   */
-  public override async getRecentGames(tag: string, user: User | null = null) {
-    const [formattedTag, type] = this.parseTag(tag);
-    const input = await this.resolveTag(formattedTag, type, user);
-
-    return super.getRecentGames(input).catch((err) => {
-      if (!err.response || !err.response.data) throw this.unknownError();
-      const error = err.response.data as
-        | RecentGamesNotFoundException
-        | PlayerNotFoundException;
-
-      if (error.message === "player") throw this.missingPlayer(type, tag);
-
-      if (error.message === "recentGames") {
-        const { displayName } = error as RecentGamesNotFoundException;
-
-        throw new ErrorMessage(
-          (t) => t("errors.noRecentGames.title"),
-          (t) =>
-            t("errors.noRecentGames.description", {
-              displayName: this.emojiDisplayName(t, displayName),
-            })
         );
       }
