+++ packages/schemas/src/status/index.ts
@@ -9,6 +9,7 @@
 import { Field } from "#metadata";
 import { Game } from "#game";
 import { PlayerStatus } from "#player";
+import { RecentGame } from "./recent-game.js";
 import type { APIData } from "@statsify/util";
 
 export class Status {
@@ -36,10 +37,16 @@ export class Status {
   @Field({ store: { required: false } })
   public map?: string;
 
+  @Field()
+  public recentGames: RecentGame[];
+
   public constructor(data: APIData) {
     this.online = data.online;
     this.game = new Game(data.gameType ?? "LIMBO");
     this.mode = data.mode;
     this.map = data.map;
+    this.recentGames = [];
   }
 }
+
+export * from "./recent-game.js";
